requirements.txt
# Core
langchain==0.1.0
langchain-community
langchain-openai
openai>=1.30.0
anthropic>=0.29.0
tiktoken>=0.6.0
python-dotenv

# Search and parsing
duckduckgo-search>=5.3.0
requests
beautifulsoup4
pydantic>=2.0.0
PyYAML

# UIs
streamlit==1.25.0
gradio==3.40.0

# Utils & tests
tenacity
pytest

google-generativeai>=0.5.0
